+++ src/Data/Singletons/Promote.hs
@@ -610,51 +610,53 @@ promoteClause :: DClause -> PrM (DTySynEqn, ADClause)
 promoteClause (DClause pats exp) = do
   -- promoting the patterns creates variable bindings. These are passed
   -- to the function promoted the RHS
-  (types, new_vars) <- evalForPair $ mapM promotePat pats
+  ((types, pats'), new_vars) <- evalForPair $ mapAndUnzipM promotePat pats
   (ty, ann_exp) <- lambdaBind new_vars $ promoteExp exp
   all_locals <- allLocals   -- these are bound *outside* of this clause
   return ( DTySynEqn (map DVarT all_locals ++ types) ty
-         , ADClause new_vars pats ann_exp )
+         , ADClause new_vars pats' ann_exp )
 
 promoteMatch :: DMatch -> PrM (DTySynEqn, ADMatch)
 promoteMatch (DMatch pat exp) = do
   -- promoting the patterns creates variable bindings. These are passed
   -- to the function promoted the RHS
-  (ty, new_vars) <- evalForPair $ promotePat pat
+  ((ty, pat'), new_vars) <- evalForPair $ promotePat pat
   (rhs, ann_exp) <- lambdaBind new_vars $ promoteExp exp
   all_locals <- allLocals
   return $ ( DTySynEqn (map DVarT all_locals ++ [ty]) rhs
-           , ADMatch new_vars pat ann_exp)
+           , ADMatch new_vars pat' ann_exp)
 
 -- promotes a term pattern into a type pattern, accumulating bound variable names
-promotePat :: DPat -> QWithAux VarPromotions PrM DType
-promotePat (DLitPa lit) = do
-  lit' <- promoteLitPat lit
-  return lit'
+promotePat :: DPat -> QWithAux VarPromotions PrM (DType, ADPat)
+promotePat (DLitPa lit) = (, ADLitPa lit) <$> promoteLitPat lit
 promotePat (DVarPa name) = do
       -- term vars can be symbols... type vars can't!
   tyName <- mkTyName name
   addElement (name, tyName)
-  return $ DVarT tyName
+  return (DVarT tyName, ADVarPa name)
 promotePat (DConPa name pats) = do
-  types <- mapM promotePat pats
+  (types, pats') <- mapAndUnzipM promotePat pats
   let name' = unboxed_tuple_to_tuple name
-  return $ foldType (DConT name') types
+  return (foldType (DConT name') types, ADConPa name pats')
   where
     unboxed_tuple_to_tuple n
       | Just deg <- unboxedTupleNameDegree_maybe n = tupleDataName deg
       | otherwise                                  = n
 promotePat (DTildePa pat) = do
   qReportWarning "Lazy pattern converted into regular pattern in promotion"
-  promotePat pat
+  fmap ADTildePa <$> promotePat pat
 promotePat (DBangPa pat) = do
   qReportWarning "Strict pattern converted into regular pattern in promotion"
-  promotePat pat
+  fmap ADBangPa <$> promotePat pat
 promotePat (DSigPa pat ty) = do
-  promoted <- promotePat pat
+  -- We must maintain the invariant that any promoted pattern signature must
+  -- not have any wildcards in the underlying pattern.
+  -- See Note [Singling pattern signatures].
+  wildless_pat <- removeWilds pat
+  (promoted, pat') <- promotePat wildless_pat
   ki <- promoteType ty
-  return $ DSigT promoted ki
-promotePat DWildPa = return DWildCardT
+  return (DSigT promoted ki, ADSigPa promoted pat' ki)
+promotePat DWildPa = return (DWildCardT, ADWildPa)
 
 promoteExp :: DExp -> PrM (DType, ADExp)
 promoteExp (DVarE name) = fmap (, ADVarE name) $ lookupVarE name
@@ -711,7 +713,7 @@ promoteExp (DLetE decs exp) = do
 promoteExp (DSigE exp ty) = do
   (exp', ann_exp) <- promoteExp exp
   ty' <- promoteType ty
-  return (DSigT exp' ty', ADSigE ann_exp ty)
+  return (DSigT exp' ty', ADSigE exp' ann_exp ty)
 promoteExp e@(DStaticE _) = fail ("Static expressions cannot be promoted: " ++ show e)
 
 promoteLitExp :: Quasi q => Lit -> q DType